Moved addition, subtraction, etc under math.hpp
